Add database entities and service for shows
This commit includes the addition of new entities to the database for 'show', 'episode', 'user' and the respective services for managing show data. Additional dependencies required for connecting and managing the database have also been added to the package.json. The controller has been updated to retrieve show data based on an id.